+++ client/scss/components/_scroll_to_info.scss
@@ -8,7 +8,7 @@
   }
 }
 
-.scroll-to--work-info {
+.scroll-to-info--work {
   position: relative;
   transform: translateY(0);
   animation: scrollToWobble 2s ease-in-out 3 forwards;
