+++ client/scss/non-critical.scss
@@ -22,7 +22,7 @@
 @import 'components/pre';
 @import 'components/quote';
 @import 'components/search_box';
-@import 'components/scroll_to';
+@import 'components/scroll_to_info';
 @import 'components/slider';
 @import 'components/slider__gallery';
 @import 'components/slider__in-article';
