+++ client/scss/utilities/_root-scope-classes.scss
@@ -78,6 +78,10 @@
   display: flex;
 }
 
+.flex--column {
+  flex-direction: column;
+}
+
 .flex-inline {
   display: inline-flex;
 }
