+++ server/views/components/scroll-to-info/scroll-to-info.config.js
@@ -1,0 +1,18 @@
+export const display = {
+  background: '#1d1d1d'
+};
+
+export const context = {
+  model: {
+    elementId: 'work-info',
+    speed: 1000,
+    easing: 'easeInOutQuad'
+  },
+  modifiers: {
+    'work': true
+  },
+  data: {
+    extraInformationIconClasses: ['icon--white', 'icon--2x'],
+    extraChevronIconClasses: ['icon--white']
+  }
+};
