--- server/views/components/scroll-to/scroll-to.config.js
@@ -1,7 +1,0 @@
-export const context = {
-  model: {
-    elementId: 'work-info',
-    speed: 1000,
-    easing: 'easeInOutQuad'
-  }
-};
