--- server/views/components/scroll-to/scroll-to.njk
@@ -1,8 +1,0 @@
-<a
-  href="#{{ model.elementId }}"
-  class="scroll-to js-scroll-to plain-link {{ 'scroll-to' | componentClasses(modifiers) }} {% for class in data.extraClasses %}{{ class }} {% endfor %}"
-  data-speed="{{ model.speed }}"
-  data-easing="{{ model.easing }}"
-  data-element="{{ model.elementId }}">
-  {% icon 'actions/chevron-medium', null, data.extraIconClasses %}
-</a>
