+++ server/views/global/icons.config.js
@@ -16,7 +16,7 @@ function walk(dir) {
       return acc.concat(walk(dirOrFile));
     } else if (path.extname(dirOrFile) === '.svg') {
       return acc.concat({
-        extraClasses: ['icon--fill', 'icon--4x', 'icon--orange-text'],
+        extraClasses: ['icon--fill', 'icon--4x'],
         icon: path.join(path.basename(dir), path.basename(dirOrFile, '.svg'))
       });
     }
