+++ server/views/pages/work.njk
@@ -28,7 +28,7 @@
     <div class="container">
       <div class="grid">
         <div class="{{ {s:12, m:12, l:12, xl: 12} | gridClasses }}">
-          {% componentV2 'scroll-to', {elementId: 'work-info', easing: 'easeInOutQuad'}, {'work-info': true}, {extraClasses: ['float-r'], extraIconClasses: ['icon--white']} %}
+          {% componentV2 'scroll-to-info', {elementId: 'work-info', easing: 'easeInOutQuad'}, {'work': true}, {extraClasses: ['float-r'], extraInformationIconClasses: ['icon--2x', 'icon--white'], extraChevronIconClasses: ['icon--white']} %}
         </div>
       </div>
     </div>
